Refactor OverloadIntegrationTest breaking out a test base, and the fake resource monitors.
